网站备份清理规则

+++ scripts/backup.py
@@ -68,26 +68,11 @@ class backupTools:
         log = "网站[" + name + "]备份成功,用时[" + str(round(outTime, 2)) + "]秒"
         mw.writeLog('计划任务', log)
         print("★[" + endDate + "] " + log)
-        print("|---保留最新的[" + save + "]份备份")
+        
         print("|---文件名:" + filename)
-
-        # 清理多余备份
-        backups = sql.table('backup').where(
-            'type=? and pid=?', ('0', pid)).field('id,filename').select()
-
-        num = len(backups) - int(save)
-        if num > 0:
-            for backup in backups:
-                mw.execShell("rm -f " + backup['filename'])
-                sql.table('backup').where('id=?', (backup['id'],)).delete()
-                num -= 1
-                print("|---已清理过期备份文件：" + backup['filename'])
-                if num < 1:
-                    break
+        self.cleanBackup('0', pid, save)
 
     def backupDatabase(self, name, save):
-        
-        print("save", save)
         db_path = mw.getServerDir() + '/mysql-apt'
         db_name = 'mysql'
         find_name = mw.M('databases').dbPos(db_path, 'mysql').where(
@@ -147,6 +132,24 @@ class backupTools:
         log = "数据库[" + name + "]备份成功,用时[" + str(round(outTime, 2)) + "]秒"
         mw.writeLog('计划任务', log)
         print("★[" + endDate + "] " + log)
+        print("|---文件名:" + filename)
+        self.cleanBackup('1', pid, save)
+
+
+    def backupDatabaseAll(self, save):
+        db_path = mw.getServerDir() + '/mysql-apt'
+        db_name = 'mysql'
+        databases = mw.M('databases').dbPos(
+            db_path, db_name).field('name').select()
+        for database in databases:
+            self.backupDatabase(database['name'], save)
+
+    def backupSiteAll(self, save):
+        sites = mw.M('sites').field('name').select()
+        for site in sites:
+            self.backupSite(site['name'], save)
+    
+    def cleanBackup(self, type, pid, save):
         # 清理多余备份
         saveAllDay = int(save.get('saveAllDay'))
         saveOther = int(save.get('saveOther'))
@@ -154,9 +157,8 @@ class backupTools:
 
         # saveAllDay天内全部保留，其余只保留saveOther份，最长保留saveMaxDay天
         print("|---[" + str(saveAllDay) + "]天内全部保留，其余只保留[" + str(saveOther) + "]份，最长保留[" + str(saveMaxDay) + "]天")
-        print("|---文件名:" + filename)
 
-        backups = mw.M('backup').where('type=? and pid=?', ('1', pid)).field('id,filename,addtime').order('addtime desc').select()
+        backups = mw.M('backup').where('type=? and pid=?', (type, pid)).field('id,filename,addtime').order('addtime desc').select()
 
         # 获取当前日期
         now = datetime.datetime.now()
@@ -190,27 +192,15 @@ class backupTools:
             # 最长保留saveMaxDay天
             if days > saveMaxDay:
                 to_delete.extend(backups_on_date)
+        if len(to_delete) == 0:
+            print("|---没有需要清理的备份")
+            return
 
         # 删除需要删除的备份
         for backup in to_delete:
             os.system("rm -f " + backup['filename'])
             mw.M('backup').where('id=?', (backup['id'],)).delete()
             print("|---已清理过期备份文件：" + backup['filename'])
-
-
-    def backupDatabaseAll(self, save):
-        db_path = mw.getServerDir() + '/mysql-apt'
-        db_name = 'mysql'
-        databases = mw.M('databases').dbPos(
-            db_path, db_name).field('name').select()
-        for database in databases:
-            self.backupDatabase(database['name'], save)
-
-    def backupSiteAll(self, save):
-        sites = mw.M('sites').field('name').select()
-        for site in sites:
-            self.backupSite(site['name'], save)
-
 
 if __name__ == "__main__":
     backup = backupTools()
